crypto-alert-system/price-checker/requirements.txt
google-cloud-firestore==2.*
requests==2.*
functions-framework==3.*
sendgrid